+++ CHANGES.md
@@ -7,7 +7,7 @@
   `Cobertura.__init__(filesystem=...)`. If the content of files is accessed
   (e.g. to render a full coverage report) then a `FileSystem` instance must be
   provided.
-* The class `Cobertura` will raise `FileSystemMissingForSource` if no
+* The class `Cobertura` will raise `MissingFileSystem` if no
   `FileSystem` object was provided via the keyword argument
   `Cobertura(filesystem=...)`. It will only be raised when calling methods that
   attempt to read the content of files, e.g. `Cobertura.file_source(...)` or